Add Astro starter template as a submodule
This commit adds the Astro starter template as a submodule in the templates/starter directory. The submodule is sourced from https://github.com/zankhq/astro-starter.git.

Update package.json with project description

This commit updates the package.json file to include a description for the project. The description now reads "A CLI to create a new project with the AT-AT stack".

Add Astro starter template as subproject

This commit adds the Astro starter template as a subproject in the templates/starter directory. The subproject has commit hash eeae38cf2bf2a91bd2862004d09304a85b2798d0.